refactor
git-svn-id: http://biztracker.googlecode.com/svn/trunk@173 d1847493-4e3b-e4a7-23b0-afffdb14bbb4

+++ src/com/xiaolei/android/customControl/CurrencyView.java
@@ -6,6 +6,7 @@ import com.xiaolei.android.common.Utility;
 import android.R.color;
 import android.content.Context;
 import android.util.AttributeSet;
+import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.widget.LinearLayout;
@@ -60,7 +61,7 @@ public class CurrencyView extends LinearLayout {
 	}
 
 	protected void updateLayout() {
-		if (mCost != 0) {
+		if ((long)mCost != 0) {
 			mTextViewCost.setText(Utility.formatCurrency(mCost, mCurrencyCode,
 					false));
 			mTextViewCurrencyCode.setText(mCurrencyCode);
